add dispenser to state
